added MRTK object (recognise hands, but not airtaps)
MRTollkit object has been added (it has "camera" and "Spatial Awareness" features deactivated.
When built it is able to recognise the hands, but when you do airtaps, no sphere is shown. (Maybe it is because it needs the Spatial Awareness tool to be active).
Careful, having to ways of understanding the scene.