create trajectory when deviates
Now the trajectory is recalculated if the user deviates from the path
Also there are two ways to calculate path:
- Wide and traditional: when the agent moves more than a certain distance, a new point in the line renderer is added. It gives a smoother path but it takes some time to create
- Narrow, straight and new: waits until the path is generated with the setDestination method (better than the calculatePath method) and adds the corners to the line renderer.
The path is made of a few lines (It is less smooth), and it's calculated faster. It is used for calculating the deviation of the user.

NEED TO DECIDE WHICH PATH TO BE SHOWN